requirements.txt
# Core Dependencies
gradio>=4.41.0
openai>=1.40.0
PyMuPDF>=1.24.10
faiss-cpu>=1.8.0.post1
numpy>=1.26.0

# Optional Dependencies
pandas>=2.2.0
python-dotenv>=1.0.1
